extract functions out of functors
